reorg(2/5): extract resilience primitives to resilience/
Moves the HTTP-hardening primitives out of `services/` (where they
were mixed with the API clients) into a dedicated `src/resilience/`
folder. These are pure, reusable utilities with no coupling to any
particular client:

- `retry-backoff.ts`   — exponential backoff with symmetric jitter
- `retry-guard.ts`     — cooldown gate for auth-retry storms
- `rate-limit-gate.ts` — Retry-After-aware pause gate
- `session-expiry.ts`  — expiry check with malformed-input handling
- `disposable-timeout.ts` — used by RetryGuard + SyncManager
- `index.ts`           — was `services/resilience.ts` (barrel)

`resilience/index.ts` continues to re-export the four domain error
classes from `src/errors.ts` for convenience — `errors.ts` stays at
top-level because it's part of the public API surface.

`RequestLifecycleEmitter` is temporarily re-exported via a bridge
from `../services/events-emitter.ts`; commit 3 will move that file
to observability/ and remove the bridge.

This is pure file relocation — no behavior change, 460 tests passing,
100 % coverage preserved.
